+++ rts-exchange-gateway-processor/src/main/resources/db/schema/002_create_price_table.sql
@@ -1,5 +1,5 @@
 create table price (
     symbol text,
-    price bigint,
+    price numeric(18,8),
     event_time timestamptz
 )
